+++ src/unsw/loopmania/Damageable.java
@@ -1,7 +1,7 @@
 package unsw.loopmania;
 
 public interface Damageable {
-    public void takeDamage(int damage);
+    public void takeDamage(double damage);
     
     public void dealDamage(Damageable damageable);
 }
